+++ src/serializers/IgbinarySerializer.php
@@ -4,12 +4,16 @@
  * @copyright Copyright (c) 2008 Yii Software LLC
  * @license http://www.yiiframework.com/license/
  */
+
 namespace yii\queue\serializers;
 
 use yii\base\Object;
 
 /**
  * Class IgbinarySerializer
+ *
+ * It uses an alternative serializer available via PECL extension which produces
+ * more compact data chunks significantly faster that native PHP one.
  *
  * @author xutl <xutongle@gmail.com>
  */
@@ -22,6 +26,7 @@ class IgbinarySerializer extends Object implements SerializerInterface
     {
         return igbinary_serialize($job);
     }
+
     /**
      * @inheritdoc
      */